backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.0
yfinance==0.2.28
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
matplotlib==3.7.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2
python-jose[cryptography]==3.3.0
